base/requirements/base.txt
# ------------------------------------------------------------------------------
# Django
# ------------------------------------------------------------------------------
# https://www.djangoproject.com/
django==4.0.9  # pyup: < 4.1
# https://github.com/joke2k/django-environ
django-environ==0.9.0
# https://github.com/jazzband/django-model-utils
django-model-utils==4.3.1
# https://github.com/adamchainz/django-cors-headers
django-cors-headers==3.13.0
# https://github.com/carltongibson/django-filter/tree/main
django-filter==23.1
# https://www.django-rest-framework.org/
djangorestframework==3.14.0
# ------------------------------------------------------------------------------
# Extras
# ------------------------------------------------------------------------------
# https://github.com/tqdm/tqdm
tqdm==4.64.1
